Did same re-factor with attachments
